Create user component
